Update and Delete  Authorization functionality added.

+++ self_practices/assaignment_1/search_tweets/twitter/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, get_object_or_404, redirect
 from django.http import HttpResponse
 from .models import Tweet 
 from .forms import TweetForm
+from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 def tweet_list(request):
@@ -14,6 +15,7 @@ def tweet_detail(request, tweet_id):
     
     return render(request, 'tweet_detail.html', {'tweet': tweet})
 
+@login_required
 def tweet_create(request):
     if request.method == "POST":
         form = TweetForm(request.POST, request.FILES)
@@ -30,8 +32,12 @@ def tweet_create(request):
 
     return render(request, 'tweet_create.html', {'form': form})
 
+@login_required
 def tweet_update(request, tweet_id):
     tweet = get_object_or_404(Tweet, pk = tweet_id)
+
+    if tweet.user != request.user:
+        return HttpResponse('You are not allowed here!')
 
     if request.method == "POST":
         form = TweetForm(request.POST, request.FILES, instance = tweet)
@@ -48,8 +54,12 @@ def tweet_update(request, tweet_id):
 
     return render(request, 'tweet_update.html', {'form': form})
 
+@login_required
 def tweet_delete(request, tweet_id):
-    tweet = get_object_or_404(Tweet, pk = tweet_id, user = request.user)
+    tweet = get_object_or_404(Tweet, pk = tweet_id)
+
+    if tweet.user != request.user:
+         return HttpResponse('You are not allowed here!')
 
     if request.method == "POST":
         tweet.delete()
